requirements.txt
streamlit
langchain
loguru
streamlit_chat
openai
python-dotenv